Complete library finishing pass: all 8 tasks done
- Fix __init__.py: export 31 symbols (added viz functions, __all__)
- pyproject.toml: optional deps [viz/docs/all], strict mypy, ruff rules
- Type annotations: mypy strict clean (0 errors, 16 source files)
- Sphinx docs: build clean with -W (warnings as errors)
- Pre-commit config: ruff, mypy, standard hooks
- Examples: autograd_intro, black_scholes_greeks, tt_compression_demo
- Integration tests: 21 tests (vectorized Greeks, validation, put-call
  parity, top-level imports, surrogate edge cases)
- Wheel build verified: sdist + wheel, all 31 symbols importable
- .gitignore: exclude docs/_build/

Total: 207 tests passing, mypy clean, Sphinx clean.

+++ src/tensorquantlib/tt/ops.py
@@ -384,4 +384,4 @@ def tt_frobenius_norm(cores: List[np.ndarray]) -> float:
     Returns:
         Frobenius norm (scalar).
     """
-    return np.sqrt(max(0.0, tt_dot(cores, cores)))
+    return float(np.sqrt(max(0.0, tt_dot(cores, cores))))
